+++ src/components/Auth/index.ts
@@ -1,0 +1,4 @@
+export * from './ForgotPassword';
+export * from './Login';
+export * from './ResendAuthEmail';
+export * from './ChangePassword';
