+++ src/containers/Auth/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Route, Switch, useRouteMatch } from 'react-router-dom';
-import Login from '../components/Auth/Login';
-import ForgotPassword from '../components/Auth/ForgotPassword';
-import ResetPassword from '../components/Auth/ResetPassword';
+import Login from '../../components/Auth/Login';
+import ForgotPassword from '../../components/Auth/ForgotPassword';
+import { ResetPassword } from './ResetPassword';
 
 const Auth = () => {
     const { path } = useRouteMatch();
